@@ -2,17 +2,8 @@
 
 declare(strict_types=1);
 
-/*
- * (c) Jeroen van den Enden <info@endroid.nl>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
 namespace Endroid\ImagePlaceholder\Exception;
 
-use Exception;
-
-class ProviderNotFoundException extends Exception
+class ProviderNotFoundException extends \Exception
 {
 }
