@@ -2,18 +2,12 @@
 
 declare(strict_types=1);
 
-/*
- * (c) Jeroen van den Enden <info@endroid.nl>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
 namespace Endroid\ImagePlaceholder\Provider;
 
 interface ProviderInterface
 {
-    public function getUrl($width, $height, array $options = []): string;
+    /** @param array<mixed> $options */
+    public function getUrl(int $width, int $height, array $options = []): string;
 
     public function getName(): string;
 }
